requirements.txt
bcrypt~=4.0.1
fastapi~=0.101.0
fastapi_login==1.9.1
fastapi_utils==0.2.1
numpy==1.25.0
orjson==3.9.10
pandas==1.5.3
passlib==1.7.4
psycopg2==2.9.3
pydantic==1.9.1
PyJWT==2.8.0
PyJWT==2.8.0
pyspark==3.4.1
python-decouple==3.8
python-dotenv==1.0.0
# python_bcrypt==0.3.2
python-multipart~=0.0.6
Requests==2.31.0
SQLAlchemy==1.4.39
starlette==0.27.0
tqdm==4.66.1
uvicorn==0.23.2
rasterio==1.4.3
geojson==3.2.0
geojson-pydantic==1.2.0
matplotlib==3.9.4
seaborn==0.13.2
Jinja2==3.1.4
pdfkit==1.0.0
plotly==6.0.0
kaleido==0.2.1
psycopg2==2.9.3
markdown==3.0.0
contextily==1.6.2
openpyxl==3.1.5
OWSLib==0.31.0
